+++ src/main/java/org/folio/service/dataimport/IdStorageService.java
@@ -3,19 +3,15 @@ package org.folio.service.dataimport;
 import io.vertx.core.Future;
 
 /**
- * Service that manages associations between record id and entity id
+ * Service that manages record id deduplication.
  */
 public interface IdStorageService {
 
   /**
-   * Stores association between specified record id and entity id and returns specified {@code entityId}.
-   * If the specified {@code record id} already associated with entity id, then returns
-   * that existing (previously associated) entity id without saving association with the specified {@code entityId}.
-   *
+   * Stores specific recordId inside DB. If it is already exists - failed future with DuplicateEventException.
    * @param recordId - record id
-   * @param entityId - entity id
    * @param tenantId - tenant id
-   * @return future with entity id that associated with specified {@code recordId}
+   * @return future with record id  {@code recordId}
    */
-  Future<String> store(String recordId, String entityId, String tenantId);
+  Future<String> store(String recordId, String tenantId);
 }