+++ src/main/java/org/folio/service/dataimport/handlers/CreateOrderEventHandler.java
@@ -64,6 +64,7 @@ import static java.lang.Boolean.FALSE;
 import static java.lang.String.format;
 import static org.apache.commons.lang3.StringUtils.isBlank;
 import static org.apache.commons.lang3.StringUtils.isNotBlank;
+import static org.apache.commons.lang3.StringUtils.isEmpty;
 import static org.folio.ActionProfile.Action.CREATE;
 import static org.folio.ActionProfile.FolioRecord.HOLDINGS;
 import static org.folio.ActionProfile.FolioRecord.INSTANCE;
@@ -165,59 +166,74 @@ public class CreateOrderEventHandler implements EventHandler {
     Map<String, String> okapiHeaders = DataImportUtils.extractOkapiHeaders(dataImportEventPayload);
     OkapiConnectionParams okapiParams = getOkapiConnectionParams(okapiHeaders, vertx);
     String sourceRecordId = dataImportEventPayload.getContext().get(RECORD_ID_HEADER);
-    Optional<Integer> poLinesLimitOptional = extractPoLinesLimit(dataImportEventPayload);
+    idStorageService.store(sourceRecordId, dataImportEventPayload.getTenant())
+      .onComplete(result -> {
+        if (result.failed()) {
+          future.completeExceptionally(result.cause());
+        } else {
+          Optional<Integer> poLinesLimitOptional = extractPoLinesLimit(dataImportEventPayload);
+          MappingManager.map(dataImportEventPayload, new MappingContext());
 
-    RequestContext requestContext = new RequestContext(Vertx.currentContext(), okapiHeaders);
-    Future<JsonObject> tenantConfigFuture = configurationEntriesService.loadConfiguration(ORDER_CONFIG_MODULE_NAME, requestContext);
-    String temporaryOrderIdForANewOrder = UUID.randomUUID().toString();
+          RequestContext requestContext = new RequestContext(Vertx.currentContext(), okapiHeaders);
+          Future<JsonObject> tenantConfigFuture = configurationEntriesService.loadConfiguration(ORDER_CONFIG_MODULE_NAME, requestContext);
+          String temporaryOrderIdForANewOrder = UUID.randomUUID().toString();
 
-    tenantConfigFuture
-      .onSuccess(tenantConfig -> overridePoLinesLimit(tenantConfig, poLinesLimitOptional))
-      .compose(tenantConfig -> {
-        Promise<Void> promise = Promise.promise();
-        prepareEventPayloadForMapping(dataImportEventPayload);
-        mappingParametersCache.get(okapiParams)
-          .onSuccess(mappingParameters -> {
-            MappingManager.map(dataImportEventPayload, new MappingContext().withMappingParameters(mappingParameters));
-            promise.complete();
-          })
-          .onFailure(promise::fail);
-        return promise.future();
-      })
-      .compose(orderId -> prepareMappingResult(dataImportEventPayload))
-      .compose(v -> setApprovedFalseIfUserNotHaveApprovalPermission(dataImportEventPayload, tenantConfigFuture.result(), requestContext))
-      .compose(tenantConfig -> generateSequentialOrderId(dataImportEventPayload, tenantConfigFuture.result(), temporaryOrderIdForANewOrder))
-      .compose(generatedOrderId -> {
-        LOGGER.info("handle:: generatedOrderId = {}, jobExecutionId {}", generatedOrderId, dataImportEventPayload.getJobExecutionId());
-        if (temporaryOrderIdForANewOrder.equals(generatedOrderId)) {
-          LOGGER.info("handle:: new order with id: {} should be created for jobExecutionId: {}",
-            generatedOrderId, dataImportEventPayload.getJobExecutionId());
-          //TODO: the deduplication should be changed to poLines in the near future
-          return idStorageService.store(sourceRecordId, generatedOrderId, dataImportEventPayload.getTenant())
-            .compose(savedOrder -> saveOrder(dataImportEventPayload, generatedOrderId, tenantConfigFuture.result(), requestContext))
-            .map(CompositePurchaseOrder::getId);
+          tenantConfigFuture
+            .onSuccess(tenantConfig -> overridePoLinesLimit(tenantConfig, poLinesLimitOptional))
+            .compose(tenantConfig -> {
+              Promise<Void> promise = Promise.promise();
+              prepareEventPayloadForMapping(dataImportEventPayload);
+              mappingParametersCache.get(okapiParams)
+                .onSuccess(mappingParameters -> {
+                  MappingManager.map(dataImportEventPayload, new MappingContext().withMappingParameters(mappingParameters));
+                  promise.complete();
+                })
+                .onFailure(promise::fail);
+              return promise.future();
+            })
+            .compose(orderId -> prepareMappingResult(dataImportEventPayload))
+            .compose(v -> setApprovedFalseIfUserNotHaveApprovalPermission(dataImportEventPayload, tenantConfigFuture.result(), requestContext))
+            .compose(tenantConfig -> generateSequentialOrderId(dataImportEventPayload, tenantConfigFuture.result(), temporaryOrderIdForANewOrder))
+            .compose(generatedOrderId -> {
+              if (isEmpty(generatedOrderId)) {
+                String errorMessage = format("handle:: generatedOrderId is null, jobExecutionId %s", dataImportEventPayload.getJobExecutionId());
+                LOGGER.error(errorMessage);
+                return Future.failedFuture(new EventProcessingException(errorMessage));
+              }
+              LOGGER.info("handle:: generatedOrderId = {}, jobExecutionId {}", generatedOrderId, dataImportEventPayload.getJobExecutionId());
+              if (temporaryOrderIdForANewOrder.equals(generatedOrderId)) {
+                LOGGER.info("handle:: new order with id: {} should be created for jobExecutionId: {}",
+                  generatedOrderId, dataImportEventPayload.getJobExecutionId());
+                return saveOrder(dataImportEventPayload, generatedOrderId, tenantConfigFuture.result(), requestContext)
+                  .map(CompositePurchaseOrder::getId);
+              }
+              return Future.succeededFuture(generatedOrderId);
+            })
+            .compose(orderId -> checkIfOrderSaved(orderId, requestContext, dataImportEventPayload.getJobExecutionId(), vertx))
+            .compose(orderId -> saveOrderLines(orderId, dataImportEventPayload, tenantConfigFuture.result(), requestContext))
+            .compose(v -> adjustEventType(dataImportEventPayload, tenantConfigFuture.result(), okapiParams, requestContext))
+            .onComplete(ar -> {
+              if (ar.failed()) {
+                LOGGER.error("handle:: Error during order or order line creation for jobExecutionId: {}",
+                  dataImportEventPayload.getJobExecutionId(), ar.cause());
+                clearOrderIdInPoLineEntityIfNecessary(dataImportEventPayload);
+                future.completeExceptionally(ar.cause());
+                return;
+              }
+              future.complete(dataImportEventPayload);
+            });
         }
-        return Future.succeededFuture(generatedOrderId);
-      })
-      .compose(orderId -> checkIfOrderSaved(orderId, requestContext, dataImportEventPayload.getJobExecutionId(), vertx))
-      .compose(orderId -> saveOrderLines(orderId, dataImportEventPayload, tenantConfigFuture.result(), requestContext))
-      .compose(v -> adjustEventType(dataImportEventPayload, tenantConfigFuture.result(), okapiParams, requestContext))
-      .onComplete(ar -> {
-        if (ar.failed()) {
-          LOGGER.error("handle:: Error during order or order line creation for jobExecutionId: {}",
-            dataImportEventPayload.getJobExecutionId(), ar.cause());
-          clearOrderIdInPoLineEntityIfNecessary(dataImportEventPayload);
-          future.completeExceptionally(ar.cause());
-          return;
-        }
-        future.complete(dataImportEventPayload);
       });
-
     return future;
   }
 
   private Future<String> checkIfOrderSaved(String orderId, RequestContext requestContext, String jobExecutionId, Vertx vertx) {
     LOGGER.info("checkIfOrderSaved:: orderId: {}, jobExecutionId: {}", orderId, jobExecutionId);
+    if (isEmpty(orderId)) {
+      String errorMessage = format("checkIfOrderSaved:: orderId is null, jobExecutionId %s", jobExecutionId);
+      LOGGER.error(errorMessage);
+      return Future.failedFuture(new EventProcessingException(errorMessage));
+    }
     Promise<String> finalPromise = Promise.promise();
     vertx.setTimer(1000L, timerId ->
       circuitBreaker.execute(promise -> {
@@ -365,11 +381,11 @@ public class CreateOrderEventHandler implements EventHandler {
     Record parsedMarcBibRecord = new JsonObject(dataImportEventPayload.getContext().get(MARC_BIBLIOGRAPHIC.value())).mapTo(Record.class);
     if (parsedMarcBibRecord.getOrder() == null) {
       return Future.failedFuture(new IllegalArgumentException(
-        String.format("Order parameter is missing. jobExecutionId: %s", dataImportEventPayload.getJobExecutionId())));
+        String.format("generateSequentialOrderId:: Order parameter is missing. jobExecutionId: %s", dataImportEventPayload.getJobExecutionId())));
     }
 
-    int exponentOrder = parsedMarcBibRecord.getOrder() / poLinesLimit;
-    return sequentialOrderIdService.store(dataImportEventPayload.getJobExecutionId(), exponentOrder, orderId, dataImportEventPayload.getTenant());
+    int sequentialNo = parsedMarcBibRecord.getOrder() / poLinesLimit;
+    return sequentialOrderIdService.store(dataImportEventPayload.getJobExecutionId(), sequentialNo, orderId, dataImportEventPayload.getTenant());
   }
 
   private Future<CompositePurchaseOrder> saveOrder(DataImportEventPayload dataImportEventPayload, String orderId,
@@ -534,11 +550,11 @@ public class CreateOrderEventHandler implements EventHandler {
     }
 
     if (poLineJson.getJsonObject(POL_ERESOURCE_FIELD) != null
-      && !NONE.toString().equals(poLineJson.getJsonObject(POL_ERESOURCE_FIELD).getString(POL_CREATE_INVENTORY_FIELD))){
+      && !NONE.toString().equals(poLineJson.getJsonObject(POL_ERESOURCE_FIELD).getString(POL_CREATE_INVENTORY_FIELD))) {
       poLineJson.getJsonObject(POL_ERESOURCE_FIELD).put(POL_CREATE_INVENTORY_FIELD, createInventoryFieldValue.value());
     }
     if (poLineJson.getJsonObject(POL_PHYSICAL_FIELD) != null
-      && !NONE.toString().equals(poLineJson.getJsonObject(POL_PHYSICAL_FIELD).getString(POL_CREATE_INVENTORY_FIELD))){
+      && !NONE.toString().equals(poLineJson.getJsonObject(POL_PHYSICAL_FIELD).getString(POL_CREATE_INVENTORY_FIELD))) {
       poLineJson.getJsonObject(POL_PHYSICAL_FIELD).put(POL_CREATE_INVENTORY_FIELD, createInventoryFieldValue.value());
     }
     return null;