+++ src/main/resources/templates/db_scripts/create_processed_records_table.sql
@@ -1,0 +1,5 @@
+-- Creates table to store record id for deduplication
+CREATE TABLE IF NOT EXISTS processed_records (
+  record_id uuid NOT NULL PRIMARY KEY,
+  created_date timestamp
+);