+++ src/main/resources/templates/db_scripts/create_records_orders_table.sql
@@ -1,5 +1,0 @@
--- Creates table to store association between record id and order id
-CREATE TABLE IF NOT EXISTS records_orders (
-  record_id uuid NOT NULL PRIMARY KEY,
-  order_id uuid NOT NULL
-);