+++ src/test/java/org/folio/service/dataimport/handlers/CreateOrderEventHandlerTest.java
@@ -1212,6 +1212,7 @@ public class CreateOrderEventHandlerTest extends DiAbstractRestTest {
     for (int i = 0; i < recordsNumber; i++) {
       record = getRecord(i);
       dataImportEventPayload = new DataImportEventPayload()
+        .withCurrentNode(profileSnapshotWrapper.getChildSnapshotWrappers().get(0))
         .withJobExecutionId(jobExecutionJson.getString(ID_FIELD))
         .withEventType(DI_MARC_BIB_FOR_ORDER_CREATED.value())
         .withTenant(TENANT_ID)
@@ -1387,6 +1388,74 @@ public class CreateOrderEventHandlerTest extends DiAbstractRestTest {
     assertNull(poLine.getPhysical().getMaterialSupplier());
 
     assertNull(poLine.getEresource());
+  }
+
+  @Test
+  public void shouldReturnFailedByDuplicateEventExceptionFutureWhenRecordIdIsDuplicated(TestContext context) {
+    Async async = context.async();
+    ProfileSnapshotWrapper profileSnapshotWrapper = buildProfileSnapshotWrapper(jobProfile, actionProfile, mappingProfile);
+    addMockEntry(JOB_PROFILE_SNAPSHOTS_MOCK, profileSnapshotWrapper);
+
+    String recordId = UUID.randomUUID().toString();
+
+    DataImportEventPayload dataImportEventPayload = new DataImportEventPayload()
+      .withCurrentNode(profileSnapshotWrapper.getChildSnapshotWrappers().get(0))
+      .withEventType(DI_MARC_BIB_FOR_ORDER_CREATED.value())
+      .withJobExecutionId(jobExecutionJson.getString(ID_FIELD))
+      .withTenant(TENANT_ID)
+      .withOkapiUrl(OKAPI_URL)
+      .withToken(TOKEN)
+      .withContext(new HashMap<>() {{
+        put(MARC_BIBLIOGRAPHIC.value(), Json.encode(record));
+        put(JOB_PROFILE_SNAPSHOT_ID_KEY, profileSnapshotWrapper.getId());
+        put(RECORD_ID_HEADER, recordId);
+      }});
+
+    CreateOrderEventHandler createOrderHandler = getBeanFromSpringContext(vertx, CreateOrderEventHandler.class);
+    vertx.runOnContext(event -> Future.fromCompletionStage(createOrderHandler.handle(dataImportEventPayload))
+      .onComplete(e -> Future.fromCompletionStage(createOrderHandler.handle(dataImportEventPayload))
+        .onComplete(context.asyncAssertFailure(th -> {
+          context.assertTrue(th instanceof DuplicateEventException);
+          async.complete();
+        }))));
+  }
+
+  @Test
+  public void shouldNotReturnFailedByDuplicateEventExceptionFutureWhenRecordIdIsDifferent(TestContext context) {
+    Async async = context.async();
+    ProfileSnapshotWrapper profileSnapshotWrapper = buildProfileSnapshotWrapper(jobProfile, actionProfile, mappingProfile);
+    addMockEntry(JOB_PROFILE_SNAPSHOTS_MOCK, profileSnapshotWrapper);
+
+    String recordId = UUID.randomUUID().toString();
+    String newRecordId = UUID.randomUUID().toString();
+
+    DataImportEventPayload dataImportEventPayload = new DataImportEventPayload()
+      .withCurrentNode(profileSnapshotWrapper.getChildSnapshotWrappers().get(0))
+      .withEventType(DI_MARC_BIB_FOR_ORDER_CREATED.value())
+      .withJobExecutionId(jobExecutionJson.getString(ID_FIELD))
+      .withTenant(TENANT_ID)
+      .withOkapiUrl(OKAPI_URL)
+      .withToken(TOKEN)
+      .withContext(new HashMap<>() {{
+        put(MARC_BIBLIOGRAPHIC.value(), Json.encode(record));
+        put(JOB_PROFILE_SNAPSHOT_ID_KEY, profileSnapshotWrapper.getId());
+        put(RECORD_ID_HEADER, recordId);
+      }});
+
+    CreateOrderEventHandler createOrderHandler = getBeanFromSpringContext(vertx, CreateOrderEventHandler.class);
+    vertx.runOnContext(event -> Future.fromCompletionStage(createOrderHandler.handle(dataImportEventPayload))
+      .onComplete(e -> {
+        dataImportEventPayload.withContext(new HashMap<>() {{
+          put(MARC_BIBLIOGRAPHIC.value(), Json.encode(record));
+          put(JOB_PROFILE_SNAPSHOT_ID_KEY, profileSnapshotWrapper.getId());
+          put(RECORD_ID_HEADER, newRecordId);
+        }});
+        Future.fromCompletionStage(createOrderHandler.handle(dataImportEventPayload));
+      })
+      .onComplete(context.asyncAssertSuccess(th -> {
+        context.assertEquals(th.getContext().get(RECORD_ID_HEADER), newRecordId);
+        async.complete();
+      })));
   }
 
   private ProfileSnapshotWrapper buildProfileSnapshotWrapper(JobProfile jobProfile, ActionProfile actionProfile,