+++ NEWS.md
@@ -3,6 +3,7 @@
 * [MODORDERS-773](https://issues.folio.org/browse/MODORDERS-773) - Implement action profile handler for order creation
 * [MODORDERS-844](https://issues.folio.org/browse/MODORDERS-844) - Add cache to store organisations info for mapping
 * [MODORDERS-808](https://issues.folio.org/browse/MODORDERS-808) - Create "OrdersPostProcessingEventHandler" for mod-orders
+* [MODORDERS-843](https://issues.folio.org/browse/MODORDERS-843) - Adjust deduplication mechanism to store info by POLine id
 
 ## 12.5.0 Nolana R3 2022 - Released
 This release focused on upgrade RMB and module interface dependencies